requirements.txt
streamlit 
pdfplumber 
google-generativeai 
requests
python-dotenv